Make the `ensRainbowPublicConfig` field required on `EnsNodeStackInfo` data model
Also, moves the `invariant_ensRainbowSupportedLabelSetAndVersion` schema validation method to be applied against a Zod schema for the `EnsNodeStackInfo` data model
