+++ packages/ensnode-sdk/src/ensindexer/config/zod-schemas.test.ts
@@ -190,14 +190,6 @@ describe("ENSIndexer: Config", () => {
 
       it("can parse full ENSIndexerPublicConfig with label set", () => {
         const validConfig = {
-          ensRainbowPublicConfig: {
-            version: "0.32.0",
-            labelSet: {
-              labelSetId: "subgraph",
-              highestLabelSetVersion: 0,
-            },
-            recordsCount: 100,
-          },
           labelSet: {
             labelSetId: "subgraph",
             labelSetVersion: 0,
